fix(admin): split session modules for react-refresh lint rule
